Further work on ThunarShortcutsModel, add new ThunarShortcutRow widget.
This commit introduces the new ThunarShortcutRow widget which has all
the properties necessary to visualize a file, volume, mounted archive
or remote location. The widget is used in the new ThunarShortcutsView to
represent a single item.

The communication between ThunarShortcutsModel and ThunarShortcutsView
has been improved and ThunarShortcutsView now creates expanders for all
categories and is able to insert new rows inside the correct expander.