Partially implement shortcut row activation.
Files and volumes can now be resolved/mounted and are opened properly in
the folder view. However, canceling does not work yet and, since we use
GFiles in ThunarShortcutRow, we still have to load synchronously the
ThunarFile for it before we can resolve the file with ThunarBrowser.
This needs to be improved.

Also, the selection is lost as soon as an item is clicked and a file in
the folder view is selected. This may not be as easy to solve as with a
tree view but we'll see.