Add a ThunarShortcut::changed signal for lazy loading shortcuts.
This doesn't do anything yet but we can use it to watch volumes, mounts
and regular files for changes and forward those to the model, then to
the view and then to the individual rows. Another use is to resolve URIs
into GFiles into ThunarFiles asynchronously step-by-step and emit a
"changed" signal after each transition.

Maybe this will help fixing the startup delay issues due to all the slow
GVfs D-Bus services (e.g. network).