Implement lazy loading of regular and network file shortcuts.
File shortcuts are now created with a GFile that is asynchronously
resolved into a ThunarFile in the background. The ThunarShortcut and its
corresponding ThunarShortcutRow are updated as soon as ThunarFile is
available.

This required several changes to the model, the synchronization between
shortcuts, the model and the rows as well as additional properties for
custom names to override file display names etc.

The next step will be to create file shortcuts with URIs only and
resolve those into GFiles before those are resolved into ThunarFiles.
This will perhaps fix the startup delay issues with current versions of
Thunar.