Finish implementing the mount/unmount/disconnect context menu actions.
Mounts only have a "Disconnect" item that unmounts or ejects the mount
and makes it disappear from the side pane.

Mounted volumes have an "Unmount" context menu item that can be used
to unmount the volume but have it remain visible in the shortcuts pane
(unlike in current versions of Thunar). Volumes also have a "Disconnect"
icon if, and only if, they or their mounts can be ejected (there's a
difference between unmount and eject in the GIO API, so that's why).

The eject button will always eject and hide the entire mount or device.

I am not sure this will remain the way it is now but for now this shall
do.