Wrap expanders in an event box and implement flashing on new shortcuts.
This is a first version of how the header of a collapsed category could
change it's background color several times in a row to make the user
aware when a new shortcut (typically a mount or a volume) has become
available. Alternative flashing techniques (e.g. fading) are possible
but not implemented yet.