Squashme: Initial work on getting rid of ThunarShortcutsModel.
This commit is the first in a series of changes that merge
ThunarShortcut and ThunarShortcutRow into a single widget. A new widget
ThunarShortcutGroup is added to represent the devices, places and
network shortcut categories and manage the individual ThunarShortcut
widgets added for each shortcut. All in all, this will help in getting
rid of code duplication, synchronization of asynchronously resolved file
information etc.