Initial work on adding custom and other actions to the file menu.
Adding more items to the file menu would allow users to define keyboard
shortcuts for them. Unfortunately, the action names generated for menu
items from plugins are persistent only as long as the order does not
change. We need to find a better unique name for them, so that adding
them with gtk_action_group_add_action_with_accel() generates a unique
AND persistent accel path.