Squashme: Initial work on DND in ThunarShortcut.
Already implemented is the typical DND data negotiation code. Shortcuts
with a ThunarFile already compute and suggest possible drop actions back
to GDK.

What is missing is the same for not-yet-mounted volumes, mounted
volumes and mounts. Not-yet-mounted volumes and basically everything
that has not finished resolving need to suggest all possible actions.
Mounted volumes and mounts need to properly set the ThunarFile
independently of the DND changes.

We might have to delay asking what to do, performing the drop action,
and calling drag_leave until resolving has finished after a drop. This
might be tricky as the DND callbacks are synchronous. In any case, care
needs to be taken of not destroying the drop file list too early when
delaying the actual drop action.