Drop the cancellable in ThunarShortcut. Hide as soon as ejected.
Cancelling operations is very hard to get right, so we will not
implement it for now. We didn't support it before, so this is no
regression.

Make the action button insensitive whenever it is spinning.

Hide the shortcut as soon as ejecting was successful to avoid users
clicking on a shortcut for a no-longer-present mount or volume.