Continue reordering via DND. It sort of works but is unfinished.
We should probably drop the GtkVBox for the shortcuts in a
ThunarShortcutGroup and just implement the layout mechanism inside the
group. That would allow us to fake the empty spot indicator instead of
using a real GtkImage widget. The empty spot indicator is cool as it
delivers a feeling of direct manipulation. However, using a real widget
causes a mix of ThunarShortcut and GtkImage widgets inside the vertical
box of the group, which is kind of ugly. We should just render the
indicator but not use a real widget.

Other than that, reordering via DND works now. Shortcuts only
temporarily remain in their new positions until the ThunarShortcutsView
is destroyed. Windows are out of sync.